# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.BracketLikeResponseDoc:
    properties:
      bracket_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  controllers.BracketLikesListResponse:
    properties:
      response:
        items:
          $ref: '#/definitions/controllers.BracketLikeResponseDoc'
        type: array
      status:
        type: integer
    type: object
...
  controllers.ErrorResponse:
    properties:
      error: {}
    type: object
...
  controllers.MatchupVoteResponseDoc:
    properties:
      created_at:
        type: string
      id:
        type: integer
      matchup_id:
        type: integer
      matchup_item_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  controllers.MatchupVotesListResponse:
    properties:
      response:
        items:
          $ref: '#/definitions/controllers.MatchupVoteResponseDoc'
        type: array
      status:
        type: integer
    type: object
...
  controllers.PopularBracket:
    properties:
      author_id:
        type: integer
      current_round:
        type: integer
      engagement_score:
        type: number
      id:
        type: integer
      matchup_count:
        type: integer
      rank:
        type: integer
      title:
        type: string
    type: object
  controllers.PopularBracketsEnvelope:
    properties:
      response:
        items:
          $ref: '#/definitions/controllers.PopularBracket'
        type: array
      status:
        type: string
    type: object
...
  controllers.SimpleMessageResponse:
    properties:
      response: {}
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for brackets, matchups, voting, and engagement
  title: Matchup API
  version: "1.0"
paths:
...
  /brackets/{id}/likes:
    delete:
      description: Remove the authenticated user's like from a bracket
      parameters:
      - description: Bracket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SimpleMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unlike a bracket
      tags:
      - bracket-likes
    get:
      description: Get likes for a bracket
      parameters:
      - description: Bracket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.BracketLikesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List bracket likes
      tags:
      - bracket-likes
    post:
      description: Like a bracket as the authenticated user
      parameters:
      - description: Bracket ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.SimpleMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Like a bracket
      tags:
      - bracket-likes
...
  /brackets/popular:
    get:
      description: Get top engaged brackets
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.PopularBracketsEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get popular brackets
      tags:
      - popular
...
  /users/{id}/bracket_likes:
    get:
      description: Get bracket likes for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.BracketLikesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List user bracket likes
      tags:
      - bracket-likes
...
  /users/{id}/matchup_votes:
    get:
      description: Get votes made by a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MatchupVotesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List user matchup votes
      tags:
      - votes
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Provide a valid JWT as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"